requirements.txt
django==3.1
django-ckeditor
django-widget-tweaks
Pillow
